new code organization and change in readme - Close Issue 2
