+++ .gitignore
@@ -1,4 +1,4 @@
 target
-
+justfile
 # Local files
 TODO.md
